refact: add benchmark dir
