+++ benchmark/src/main/resources/conf.properties
@@ -1,0 +1,17 @@
+ZK_CLUSTER=127.0.0.1:32200
+ZK_PATH=/udf_test
+
+WINDOW_NUM=2
+WINDOW_SIZE=1000
+JOIN_NUM=2
+
+PK_NUM=10
+PK_MAX=1000000
+TS_BASE=1652321552000
+PK_BASE=1000000
+
+# the below conf need not modified
+
+DATABASE=dl
+TABLE=t1
+DEPLOY_NAME=demo
