+++ java/openmldb-jmh/src/main/resources/conf.properties
@@ -1,21 +1,6 @@
 ZK_CLUSTER=127.0.0.1:32200
 ZK_PATH=/udf_test
 
-WINDOW_NUM=2
-WINDOW_SIZE=1000
-JOIN_NUM=2
-
-PK_NUM=10
-PK_MAX=1000000
-TS_BASE=1652321552000
-PK_BASE=1000000
-
-# the below conf need not modified
-
-DATABASE=dl
-TABLE=t1
-DEPLOY_NAME=demo
-
 MODE=request
 BATCH_SIZE=50
 TIME_DIFF=1
@@ -28,6 +13,7 @@ METHOD=insert,query
 PUT_THREAD_NUM=1
 QUERY_THREAD_NUM=1
 NEED_CREATE=false
+PK_NUM=50
 #request:batch_request
 REQUEST_RATIO=1.0
 PROCEDURE_RATIO=0.5
